Add files via upload
"FordhamProcessor.py" is used by our university only to clean and prepare the data for analysis. The data input should meet some requirements before initializing the tool. --> TODO: A document explaining the requirements for the input data will be uploaded soon.
"CreateTables.py" is necessary so that the tool can perform efficiently without reloading every time it is used (wasting about 20 minutes). A user should run this file if it is the first time they will use this tool OR if there were any changes made to the data input file. Otherwise, the user can run the tool without any pre-processing.
"Tool.py" is the actual data analysis tool. It performs the calculations and delivers conclusions about the data.
--> TODO: It still needs to be updated so that the user interface is more intuitive, include a notification that a file was saved in the current folder, and save each terminal session into a .txt file so that the results are outputted to the user are saved for future reference.